+++ zarrs_metadata/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - **Breaking**: Rename `DataTypeMetadataV3::Binary` to `Bytes` for compatibility with `zarr-python`
 
+### Removed
+- **Breaking**: Remove the `v3::array::codec::vlen_v2` module and all associated types
+
 ## [0.2.0] - 2024-11-15
 
 ### Added